package.json: fix main reference
'main' contains the name of the file to load when a module required by the package name. Current value 'pluggable.js' points to nowhere - there is no 'pluggable.js' in the package root. Point 'main' to 'dist/pluggable.js' instead so that when 'pluggable.js' is required with no specific path 'dist/pluggable.js' is required instead.